Refactor Finish Assignment and Precision Change
Fixed score on finishAssignment to be integers

+++ src/main/java/com/mango/prjmango/ui/activities/FinishAssignmentView.java
@@ -92,7 +92,7 @@ public class FinishAssignmentView extends JPanel {
                 Fonts.SEGOE_UI_18.getFont(),
                 Colors.LIGHT_GREY);
         changeScoreLabel = Components.JLabel(
-                totalCorrect + "/" + totalQuestions,
+                (int)totalCorrect + "/" + (int)totalQuestions,
                 Fonts.SEGOE_UI_18.getFont(),
                 Colors.LIGHT_GREY);
